requirements.txt
gevent==21.1.2
flask==1.1.2
Flask-SQLAlchemy==2.4.4
flask-restful==0.3.8
redis==3.5.3
bcrypt==3.2.0
python-dotenv==0.15.0
psycopg2-binary==2.8.6
python-secrets==20.11.0
flasgger==0.9.5
marshmallow_sqlalchemy==0.24.2
flask-marshmallow==0.14.0
pydantic==1.7.3
user_agents==2.2.0
pytest==6.2.2

